correct main and renderComments
