+++ api.js
@@ -1,10 +1,6 @@
 const host = "https://wedev-api.sky.pro/api/v2/lana-olhowko/comments";
 const userURL = "https://wedev-api.sky.pro/api/user";
 
-const containerPreloader = document.getElementById('container-preloader');
-
-containerPreloader.textContent = 'Пожалуйста подождите, загружаю комментарии...';
-
 export let token;
 
 export const setToken = (newToken) => {
@@ -12,6 +8,9 @@ export const setToken = (newToken) => {
 }
 
 export function getComments() {
+
+  // const containerPreloader = document.getElementById('container-preloader');
+
   return fetch(host, {
     method: "GET",
     headers: {
@@ -19,7 +18,12 @@ export function getComments() {
     }
   })
     .then((response) => {
-      containerPreloader.textContent = '';
+
+      if (response.status === 401) {
+        throw new Error("Нет авторизации");
+      }
+
+      // containerPreloader.textContent = '';
       return response.json();
     });
 }
