+++ index.html
@@ -9,12 +9,12 @@
 
 
 <body>
-    <div class="container">
+    <!-- <div class="container">
         <div id="container-preloader"></div>
 
         <ul id="list" class="comments">
 
-            <!-- Список рендерится из JS -->
+            Список рендерится из JS
 
         </ul>
 
@@ -31,7 +31,9 @@
         <div class="add-form-row">
             <button id="del-comment" class="add-form-button">Удалить последний комментарий</button>
         </div>
-    </div>
+    </div> -->
+
+<div id="app"></div>
 
     <script type="module" src="./main.js"></script>
