+++ needtodo
@@ -1,0 +1,4 @@
+условие "тернарный оператор", если токен есть ? показываем форму : (иначе) ссылку
+отрисовка страницы логина (форму авторизации) (новый файл, новая рендерфункции новой формфы)
+
+подвязать логику авторизации. (создать запрос на апишку)
